# api/swagger.yaml
basePath: /
definitions:
...
  http.changePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      version:
        type: integer
    type: object
...
  http.registerRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  http.updateProfileRequest:
    properties:
      display_name:
        type: string
      version:
        type: integer
    type: object
  http.userResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
  description: Personal Financial Manager REST API — double-entry accounting for households.
  title: PFM-Go API
  version: "1.0"
paths:
...
  /api/v1/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a user
      tags:
      - users
    get:
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Profile update input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /api/v1/users/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Password change input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (obtained from POST /auth/login).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"